revise API Key Module overview
