+++ www/packages/docs-ui/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from "./use-active-on-scroll"
+export * from "./use-child-docs"
 export * from "./use-click-outside"
 export * from "./use-collapsible"
 export * from "./use-collapsible-code-lines"
